fix: the stopping bounds are compared in the dynamics' float precision
r309 did not take off: the production node refused its configuration.
The vertical law of d25dceb4 sets the guaranteed vertical stopping
deceleration and the planned vertical acceleration both to 1.4 m/s^2, and
the validity rules compared the double parameter to the float dynamics
field widened back to double. 1.4 has no exact binary image, so the float
image sits below the double and the rule "bound <= acceleration" failed
on equality; 2.0 and 4.0 are exact and never exposed it. The rules in the
production config and in the MPPI benchmark config now cast the bound to
float first, the precision the dynamics run in. A config test loads
1.4/1.4 and expects the configuration to be valid.

+++ drone_city_nav/src/mppi/mppi_reference.cpp
@@ -142,9 +142,13 @@ bool benchmarkConfigIsValid(const BenchmarkConfig& config) noexcept {
          config.altitude_envelope.guaranteed_vertical_deceleration_mps2 <=
              config.dynamics.maximum_vertical_acceleration_mps2 &&
          stoppingCapabilityIsValid(config.stopping_capability) &&
-         config.stopping_capability.maximum_commanded_horizontal_deceleration_mps2 <=
+         // Compared in the dynamics' float precision: a bound equal to its
+         // acceleration with no exact binary image (1.4 m/s^2) is not above it.
+         static_cast<float>(config.stopping_capability
+                                .maximum_commanded_horizontal_deceleration_mps2) <=
              config.dynamics.maximum_horizontal_acceleration_mps2 &&
-         config.stopping_capability.guaranteed_vertical_deceleration_mps2 <=
+         static_cast<float>(
+             config.stopping_capability.guaranteed_vertical_deceleration_mps2) <=
              config.dynamics.maximum_vertical_acceleration_mps2 &&
          std::isfinite(config.dynamics.maximum_translational_speed_mps) &&
          config.dynamics.maximum_translational_speed_mps > 0.0F &&
